+++ src/lib.rs
@@ -21,7 +21,7 @@ pub mod utils;
 pub struct Settings {
     pub port: u16,
     pub ip: IpAddr,
-    pub req_piece_len: usize,
+    pub req_piece_len: u32,
 }
 
 pub fn run(filename: &str, settings: Settings) -> BitterResult<()> {
